requirements.txt
# python 3.8
flatland-rl
torch
tensorboard
torch_geometric
wandb